🚚 (1134): Deplacement des donnees des courrier de reponse dans 'entites'
