+++ src/infra/sequelize/queries/modificationRequest/index.ts
@@ -1,4 +1,3 @@
-export * from './getDonnéesCourriersRéponse'
 export * from './getModificationRequestAuthority'
 export * from './getModificationRequestDataForResponseTemplate'
 export * from './getModificationRequestDetails'
