--- src/modules/modificationRequest/queries/GetDonnéesCourriersRéponse.ts
@@ -1,6 +1,0 @@
-import { DonnéesCourriersRéponse, ProjectAppelOffre } from '@entities/appelOffre'
-
-export type GetDonnéesCourriersRéponse = (
-  cahierDesChargesActuel: string,
-  projectAppelOffre: ProjectAppelOffre
-) => DonnéesCourriersRéponse
